Separated the tests for StructureFunctionLightCurve and LightCurve. Fixed a bug in LightCurve! Fixed import error in notebook.
